add 增加v2版本 cursor实现

+++ miaosha-web/src/main/java/cn/monitor4all/miaoshaweb/receiver/OrderMqReceiver.java
@@ -29,7 +29,7 @@ public class OrderMqReceiver {
             // 创建订单
 //            orderService.createOrderByMq(jsonObject.getInteger("sid"),jsonObject.getInteger("userId"));
 
-            orderService.createOrderByMq(jsonObject.getString("orderId"), jsonObject.getInteger("sid"),jsonObject.getInteger("userId"));
+            orderService.createOrderByMq(jsonObject.getString("orderId"), jsonObject.getInteger("sid"), jsonObject.getLong("userId"));
             // TODO 订单创建成功后，修改 order_message表记录
 
         } catch (Exception e) {
